Real Time Payment Score/requirements.txt
kafka-python
pandas
numpy
scikit-learn
joblib
mysql-connector-python
